requirements.txt
confluent-kafka==2.9.0
psycopg2-binary==2.9.10
pydantic==2.11.3
python-dateutil==2.9.0